authors of the simple shell repo
