+++ AUTHORS
@@ -1,0 +1,4 @@
+#This file lists all contributors to the repository.
+
+Samisaac Mburu <samisaacmbugua@gmail.com>
+Emmanuel Mwaura <manuchoadero@gmail.com>
